initial version of "places" done

+++ src/app/places/places.component.ts
@@ -1,6 +1,7 @@
 import {Component, OnInit} from '@angular/core';
 import {Place} from "./place";
 import {PlacesService} from "./places.service";
+import {ConfirmDialogService} from "../components/confirmation-dialog/confirmation-dialog.service";
 
 @Component({
   selector: 'app-places',
@@ -9,13 +10,12 @@ import {PlacesService} from "./places.service";
 })
 export class PlacesComponent implements OnInit {
 
-  searchByTitle: string = '';
   selectedPlace?: Place;
   places: Place[] = [];
   displayPlaces: Place[] = [];
   submitted = false;
 
-  constructor(private placesService: PlacesService) {
+  constructor(private placesService: PlacesService, private confirmation: ConfirmDialogService) {
   }
 
   ngOnInit(): void {
@@ -30,27 +30,18 @@ export class PlacesComponent implements OnInit {
       });
   }
 
-  onSelect(place: Place): void {
+  onSelect = (place?: Place): void => {
     this.selectedPlace = place;
     this.submitted = false;
-  }
+  };
 
-  onSearchChangeEvent(event: any){
-    if (event && event.target && event.target.value) {
-      this.onSearchChange(event.target.value);
-    } else {
-      this.onSearchChange();
-    }
-  }
-
-  onSearchChange(title: string = ''){
-    this.searchByTitle = title;
+  onSearchChange = (title: string = ''): void => {
     this.displayPlaces = this.places.filter((place: Place) => {
-      return place.title && place.title.toLowerCase().includes(this.searchByTitle.toLowerCase());
+      return place.title && place.title.toLowerCase().includes(title.toLowerCase());
     });
-  }
+  };
 
-  onSubmit() {
+  onSubmit = () : void => {
     if (!this.selectedPlace) {
       return;
     }
@@ -60,23 +51,29 @@ export class PlacesComponent implements OnInit {
         this.submitted = true;
         this.getAllPlaces();
       });
-  }
+  };
 
   newPlace() {
     this.selectedPlace = new Place();
     this.submitted = false;
   }
 
-  deletePlace() {
+  deletePlace = (): void => {
     if (!this.selectedPlace || !this.selectedPlace.id) {
       this.selectedPlace = undefined;
       return;
     }
-    this.placesService.deletePlace(this.selectedPlace)
-      .subscribe((data: any) => {
-        this.selectedPlace = undefined;
-        this.submitted = true;
-        this.getAllPlaces();
+
+    this.confirmation.openConfirmation("Are you sure?", "Do you want to delete \"" + this.selectedPlace.title + "\"?")
+      .then(result => {
+        if (result && this.selectedPlace && this.selectedPlace.id) {
+          this.placesService.deletePlace(this.selectedPlace)
+            .subscribe((data: any) => {
+              this.selectedPlace = undefined;
+              this.submitted = true;
+              this.getAllPlaces();
+            });
+        }
       });
-  }
+  };
 }
